+++ src/Contracts/Object/ResourceIdentifierCollectionInterface.php
@@ -25,6 +25,7 @@ use IteratorAggregate;
  * Interface ResourceIdentifierCollectionInterface
  *
  * @package CloudCreativity\LaravelJsonApi
+ * @deprecated 2.0.0
  */
 interface ResourceIdentifierCollectionInterface extends IteratorAggregate, Countable
 {