+++ src/Http/Controllers/JsonApiController.php
@@ -298,7 +298,7 @@ class JsonApiController extends Controller
 
         $record = $store->createRecord(
             $request->getResourceType(),
-            $request->getDocument()->getResource(),
+            $request->all(),
             $request->getParameters()
         );
 
@@ -323,7 +323,7 @@ class JsonApiController extends Controller
 
         $record = $store->updateRecord(
             $request->getRecord(),
-            $request->getDocument()->getResource(),
+            $request->all(),
             $request->getParameters()
         );
 
@@ -371,7 +371,7 @@ class JsonApiController extends Controller
         $record = $store->replaceRelationship(
             $record,
             $field,
-            $request->getDocument()->getRelationship(),
+            $request->all(),
             $request->getParameters()
         );
 
@@ -397,7 +397,7 @@ class JsonApiController extends Controller
         $record = $store->addToRelationship(
             $record,
             $field,
-            $request->getDocument()->getRelationship(),
+            $request->all(),
             $request->getParameters()
         );
 
@@ -423,7 +423,7 @@ class JsonApiController extends Controller
         $record = $store->removeFromRelationship(
             $record,
             $field,
-            $request->getDocument()->getRelationship(),
+            $request->all(),
             $request->getParameters()
         );